First attempt at formally documenting available config settings.
As TOML allows comments I made it into a TOML file that can double as
a template config that someone can just drop in and get started, killing
two birds with one stone.

I'm not sure I got all the settings but it's a start.

docs: Added config settings to documentation.