cli: commit working copy before recovering from stale state
If the working copy is dirty and stale, new commit will be created on top
of the stale parent, and the working copy will point to the created commit,
not to the commit which the repo view previously pointed.

I'm not pretty sure if this is @martinvonz suggested, so please let me know
if it's wrong.