Prevent rebasing onto both commits and their ancestors
Previously, something like `rebase -r @ -d a -d a--` created a commit
that has one parent be an ancestor of another parent. Now, such
ancestors are automatically removed from the list.

The implementation simply moves the check that I introduced in
67738a87542dbaa1ab0ce0460ebcbddf4d53c7fe from `rebase_revision`
to the more low-level `rebase_commit`.

There are two potential issues to resolve still:

- The user should probably be informed when this occurs, similarly
to b21a123bc8947e51f84d498e78c952d8511a25bc (which deals with a
special case of this situation).

- I'm not sure if the use of `unwrap` in `rebase_commit` is kosher,
but I was not sure how feasible it is to make the return type of
`rebase_commit` be a `Result`.