testing: don't have cargo install test tools
I noticed that `cargo install --path .` will install the new
`fake-editor` and `fake-diff-editor` binaries, which we don't want. It
doesn't seem to work to put them in a separate crate and have a
dev-dependency on it because `cargo` doesn't like dev-dependencies to
binary targets. However, it seems that we can put them in `examples/`
to be able to access them without having `cargo install` install them.