cargo: downgrade off of yanked version of `textwrap`
I tried to upgrade off of the yanked version, i.e. 0.15.1 ->
0.16.0. However, we also have a transitive dependency on `textwrap`
via `criterion` -> `clap` -> `textwrap`. The `clap` version used by
`criterion` is `^3.1` (our own binary uses clap 4). The matching
version we currently have in our `Cargo.lock` is 3.2.22, which depends
on the yanked `textwrap` version with `^0.15.1`, so we can't upgrade
or downgrade it. This patch therefore also downgrades `clap` from
3.2.22 to 3.2.21, since that version depends on textwrap
`^0.15.0`. With that change made, we can downgrade `textwrap` to
`0.15.0`.