diff summary: add a newline before the summary to help hard-wrapping tools
Without this change, using VSCode to hard-wrap the commit message does the wrong
thing and rewraps the `JJ: ` lines into the body of the draft log message. I
haven't tested vim or emacs hard-wrapping logic but my memory of both of those
is that it would behave mostly the same. An extra newline here shouldn't hurt
anyone, so let's just add it.