build: don't use vendored OpenSSL by default
I added support for using a vendored OpenSSL in dbadbd68c022. That was
in order to get the musl binary to work. However, it shouldn't be
needed on most platforms, and we've had a few reports of issues caused
by it. Let's disable it by default and enable it specifically when
building the musl binary instead.