Make the `??` for conflicted branches red
This demonstrates the FormattedString functionality. This commit allows
much fancier coloring, but I haven't decided on a color scheme that's
an improvement. We should also avoid creating a color soup.

It's also possible to make the entire name of the divergent branch red,
but it seems unnecessary. Magenta + red looks good.