revset: wrap implementation in Rc, so it can be shared
I'm about to add `Revset::intersect()`, and I'd like that to be able
to share the iterator state from the input revsets without consuming
them. This commit prepares for that by making the internal
`evaluate()` function return a `RevsetImpl` wrapped in an `Rc`. To
preserve the `Rc`-ness while hiding it from the caller, I added a
wrapper type that implements the `Revset` trait.