cli: add test of `jj debug reindex`, do full reindexing
I broke the commands in a27da7d8d56d and thought I just fixed it in
c7cf914694a8. However, as I added a test, I realized that I made it
only reindex the commits since the previous operation. I meant for the
command to do a full reindexing of th repo. This fixes that.