git import/export: make "git_last_seen_refs" backwards-compatible
If it doesn't exist, generate it from git refs as jj did
before. This should work as long as the last operation wasn't
"undo" or "restore".

This is a temporary behavior designed for people upgrading
from older jj. It may also be useful if the file gets corrupted
by concurrent operations that ignore the lock (e.g. Dropbox).
It's unclear if it should be kept or removed.