git export: use a file to track git branches' last seen state, fix #922
This commit essentially reverts 8a440d8 AKA https://github.com/martinvonz/jj/pull/832/commits/4dfd7651798d5cafcdb66e86ab064d6e7dd01ce5

The problem that commit is that the Git repo is, unaffected by `jj undo`. So,
it is important for JJ's view of the git refs to be unaffected by `jj undo`.

Fixes #922. Also creates a bug that is fixed in a follow-up commit