sign: read commit signatures from git
Just extract the signature to be later (somehow) shown by the templater, for
example, and used otherwise.
This is slightly not pretty because the sig field in backend::Commit is only
used when reading commits, since the value of the signature directly depends
on the commit object text and the signing will happen dynamically when writing
commits.