sign: A basic CLI signing prototype with gpgme
This just adds the `-S` option, a bunch of unwraps and otherwise unhandled
errors, and no configs whatsoever.
Also there's this invasive signer argument in `start_transaction` and a few
other places.
But the good news - it actually works!
I'm going to run `cargo run -- describe -r @- -S` so that this commit actually has
the green checkmark on github :)