refactor: remove 'jj checkout' command
Some discussion on Discord basically lead to Martin deciding this would be OK;
users should generally use `jj new` instead of checkout, which is not only more
general (it can take more than one revset, in order to perform a merge) -- it's
also not trying to be something it isn't.

The idea was, people will try to do `checkout` and might feel strange if it
doesn't work. First impressions matter. But on the other hand, `git checkout` is
a particular colloquialism that even Git itself seems to be (slowly) moving away
from with `git branch` and the experimental `git switch`. Just trying to have
a familiar name isn't enough if the mechanics aren't the same, and may lead to
*more* confusion due to behavioral changes.

This can of course always be revisited later if people really *do* clamour for
that wonderful `checkout` command, but in such a case, it's *probably* better
as an alias of `new` anyway, rather than having similar-but-more-restricted
semantics.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Iwmotxnmmopxptrvzqzpsytvkyvwlrrll