refactor(working_copy): create directory on disk to mark the presence of Git submodules
In the long-term, we would of course want to actually populate and query the status of Git submodules, but, in this commit, we just make a directory (if it doesn't already exist) to represent that the Git submodule was checked out on disk.

This is to support a later commit updating the filesystem traversal to use an additive rather than subtractive approach. That is, we start from an empty tree and traverse the repository and update the tree object with the files we find on disk. When encountering a Git submodule, we need to skip traversal and use the existing tree value for it. This is the neatest approach because, in the future, when we have to actually handle Git submodules, we'll just fill out the stubs in this commit to properly handle the Git submodule.