refactor(working_copy): have `update_file_state` query its own metadata
This lets us simplify the control flow later when we want the update loop to be able to accept files as top-level work items rather than just directories. It also is equally efficient on Unix as the previous version because querying for the metadata during directory content enumeration is implemented as a separate call to fetch the metadata (but not so on Windows).