protobuf: add `--experimental_allow_proto3_optional`
We need this flag to use `optional` fields as described in https://protobuf.dev/programming-guides/proto3/#field-labels.

Note that in protobuf3, all fields are optional. If a field is not present, its default value is used: https://protobuf.dev/programming-guides/proto3/#default. Adding the `optional` descriptor merely lets us query for whether the field is absent and process it ourselves. In `prost`, this also means that the field type becomes `Option<T>` instead of plain `T`.

There are several additional complications for `optional` fields that make it worth examining the documentation carefully (such as "For message fields, the field is not set. Its exact value is language-dependent." and "Also note that if a scalar message field is set to its default, the value will not be serialized on the wire.").