refactor(working_copy): use additive rather than subtractive filesystem traversal
Before this commit, we would deal with the entire `file_states` structure as part of conducting the repo snapshot. This won't scale for large repositories, particularly when using a filesystem monitor, because we want to ensure our operations are O(changes) instead of O(repo). This commit adjusts the traversal logic to start with an empty tree rather than the previous tree.