cargo: specify MSRV in rust-toolchain.toml
Summary: Split out from from #1851. This specifies the current project-wide MSRV
in a `rust-toolchain.toml` file which — if installed — will cause 'rustup' to
download and provision the correct version of the rust compiler.

The intent is to ensure most developers stay on the same page and people doing
occasional development don't accidentally trip over themselves and submit
something based on their own toolchain i.e. newer or older features than our
MSRV.

This also migrates the Nix Flake to use `rust-toolchain.toml` and so it creates
a single source of truth for both `nix develop` and `rustup`.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I56e48d399e541d7700c7e9652ac88c57