vscode: do not autoformat markdown or toml files
Summary: This caught me several times and several instances of autoformatting
have slipped by in some of my commits; generally this is a PITA when you're
doing something like modifying markdown files and Visual Studio Code "helpfully"
autoformats the entire document for a 1-byte change, creating a 100+ line diff
out of thin air.

This is of course a function of the specified users autosave configuration and
enabled extensions. I can't possibly account for everyone's integrations or
workflows, but at least for me, I think, it's OK to turn off autoformatting here
because we aren't enforcing a consistent one anyway.

I do think that adding autoformatting for some of these things *may* be a good
idea. But in the mean time I think this is just a helpful safeguard.

As an example of this problem, I've included a small diff to remove a bunch of
redundant whitespace from the top-level Cargo workspace file; this is exactly
the kind of autoformatting that I think isn't good unless everyone agrees on
it and it only serves to bloat diffs if someone changes it later (and no, there
isn't really a standardized format for Cargo.toml yet from what I can tell —
though, tellingly, the related style guides for Cargo.toml in the manual don't
seem to recommend things like "add 60 characters of whitespace for no reason to
align comments to column 80".)

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Iea17b611d58e65135fdc702f7b59f46c