cli: pass change IDs to ui.editor where possible
Summary: It's useful to overload `jj describe` to use a different editor,
which can format or alter commit messages before passing things off to the real
editor. However, in such a context, the `ui.editor` program has no idea what
original change ID it was invoked on.

This means that any script which wants to know the change identifier it was
invoked on has no way to obtain it.

This doesn't work in every case, but for now, fix it when using `jj describe`
and `jj commit`, at least.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Id66a0cb0bc745dc8381cbb122cf83b14