cli version: Show date as part of the version next to the commit id
The date comes from the commiter date of the commit. This is so that
it's easy to tell at a glance how old a version is.

To make it easier to get UTC date, we first obtain it as a Unix
timestamp, which should always be UTC. This will make the
interaction with Nix in a subsequent commit easier.