cli: re-enable 'watchman' support when 'packaging' is enabled
Summary: 70f6e0a enabled watchman by default in the *default* features set, but
removed it from the 'packaging' set. But features are not 'additive' by default.
When building the Nix flake from this repository, it uses --features packaging,
which no longer enables watchman, because it was removed.

The easiest solution is probably just to ensure 'packaging' is always a superset
of 'default', but I'm not sure of a way to do this.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: Ia31d5b98e0324a2b19014a3cae25672d