cargo: Change prerelease versions to end in ".1-pre" instead of ".0"
So, the prerelease version is "0.9.1-pre" instead of "0.9.0".

The main goal is to make it easier to tell from "jj version" output whether it
describes a released version or a prerelease.

According to https://semver.org/#spec-item-9, if taken as a semver, this would
be ordered after "v0.9.0" and before "v0.9.1" (if we need to have a patch
release). If "v0.9.1" was released, we could move on to "v0.9.2-pre".

Cargo docs imply that they follow this spec closely:
https://doc.rust-lang.org/cargo/reference/resolver.html#semver-compatibility

### Weirder consequences

"v0.9.1-pre.1" would follow "v0.9.1-pre" if we wanted to publish two
such versions to crate.io for some reason (I don't suggest we do this).

In theory, these are sorted alphabetically, so `v0.9.1-beta < v0.9.1-pre <
v0.9.1-rc1`. This is confusing, so we should never have `.1-alpha` or `.1-beta`
(nor probably `.1-rc1`) versions. I propose that if we ever have alpha or beta
versions, they should always be `.0` versions.