tutorial: rebase commits by commit message rather than by change ID
The tutorial currently says:

> Let's now rebase B2 directly onto A:
>
>     $ jj rebase -s puqltuttrvzp -d nuvyytnqlquo

I find it fairly confusing for the following reasons:

- The change ids are different on my machine, so it is difficult for
  me to tell which of puqltuttrvzp or nuvyytnqlquo correspond to A or
  B2. I have to scroll above in the tutorial to locate this
  information again, and I don't want to have to do this.

- You haven't explained the meaning of `-s` and `-d` yet, so I cannot
  guess from those either. (Note: using `--source` and `--destination`
  would make it easy for me to guess which is which, but the tutorial
  consistently makes the choice to use short options for now.)

Instead I propose to select the commits from their commit message /
description:

    $ jj rebase -s 'description(B2)' -d 'description(A)'

This is a tad verbose but much clearer. As a reader of the tutorial I
appreciate clarity. (I just added both `descr(x)` and `msg(x)` as
shorter alises of `description(x)` to my config file, but obviously
the tutorial cannot rely on those.)