cli version: Put `+` before build data instead of `-`
Replace `jj 0.10.0-20231027-bb8af5adc4f98ff2` with `jj 0.10.0+20231027-bb8af5adc4f98ff2`.

According to https://semver.org, the build metadata should be separated by a
`+`, not `-`. If semver rules are pedantically followed, `jj
0.10.0-20231027-bb8af5adc4f98ff2` sorts before `jj 0.10.0`, whereas the build
metadata is ignored for the purposes of ordering.

I'm not sure whether anybody else is likely to treat the full `jj version`
string as a semver this pendantically, but we can :).

I discovered this when thinking about #2258; see that PR for a few more details
about pedantic semvers.