github CI: Retry tests after a 15-min timeout
The Windows tests have been quite flaky lately,
`test_concurrent_read_write_commit` can run for 5+ hours. See e.g.
https://github.com/martinvonz/jj/actions/runs/6964394504/job/18951562247

This is not the most elegant solution, but fixing the test seems
trickier.

Normally, it takes ~3 minutes, but the timeout is 15 just to be safe.

PR #2627 checks that failed tests are not retried.

Example of this working (Windows tests pass after 2 attempts):
https://github.com/martinvonz/jj/actions/runs/6974955234/job/18981391917#step:5:1510