GitHub workflow: document how one can install old Poetry locally
This is quite minor, but it took me a few minutes to figure out
the correct command.

It might be slightly better to print this text inside the build
logs, where people will be looking for certain if the CI fails,
but I didn't immediately find a good way to do so without
complicating the config too much.