github: run the nix build with multiple cores
By default, I think we're running the Nix builds with only a single core,
while `ubuntu-latest` on GitHub apparently comes with 16GB of RAM and 4 cores.
Currently Nix clocks in as the slowest CI build, at about 7-8 minutes.

This should hopefully speed up the Nix build quite a bit and bring it up to par
with the rest of the runners, which are completing builds in 3-5 minutes.

Signed-off-by: Austin Seipp <aseipp@pobox.com>