meta: cement "Jujutsu" (呪術) as the project name
Many users over the ages have asked about the name of the project, what it
means, and how to refer to it. And occasionally we've talked about changing
the name or getting rid of it, etc.

Recently, there was also confusion about the Japanese interpretation of the name
itself, which can mean either 柔術 (martial arts) or 呪術 (magic, sorcery), each
pronounced differently depending on the exact spelling that English speakers
choose. We've never clarified this or thought about it.

Instead, let's embrace it. I've seen users tend to come away thinking or saying
things like "jj is magic". As a result, the official name in Japanese is now
unambiguously 呪術.

This also adds an FAQ section about the history of the name, the interpretation
of the Japanese word itself, and how to properly refer to it in technical
writing.

Signed-off-by: Austin Seipp <aseipp@pobox.com>