sign: Allow using test signing backend by commands
Currently the TestSigningBackend can only be used from the test gradle.
This makes it hard to build command tests for signing commands as they
are tested against the CLI interface directly.

This commit renames the TestSigningBackend struct to MockSigningBackend
and moves it to the lib module, allowing us to configure it via config
and therefore use it from command tests.