Implement advance-branches for jj commit
## Feature Description

If enabled in the user or repository settings, the local branches pointing to the
parents of the revision targeted by `jj commit` will be advanced to the newly
created commit. Support for `jj new` will be added in a future change.

This behavior can be enabled by default for all branches by setting
the following in the config.toml:

```
[experimental-advance-branches]
enabled = true
```

The default value of `experimental-advance-branches.enabled` is `false`. You
can override the behavior for specific branches by setting the value of
`experimental-advance-branches.overrides` in the config.toml:

```
[experimental-advance-branches]
overrides = ["main"]
```

Branches that have been overridden behave as if the value of
`experimental-advance-branches.enabled` is negated.


This implements feature request 2338.