github release: replace unsupported `upload-release-asset` actions
It generates numerous warnings that it will stop working once Node 16 is
unsupported. This replaces it by a seemingly equivalent action that
somebody is supporting. I tired this change on
https://github.com/ilyagr/diffedit3/, and it seems to work.

I think it would be better to switch to using
https://github.com/shogo82148/actions-upload-release-asset, which is
recommended at https://github.com/actions/upload-release-asset, but that
would be a slight change to the release process (that action creates a
release when a tag is pushed, we'd probably want it to be for tags
starting with `v`), so I'll let @martinvonz do that.

Ultimately, Martin, since you do the releases, it's entirely up to you
what we do here. For example, you can wait until the next release and
experiment with this PR then.