test: rename `coalesce(x,y)` alias in `test_templater` [NFC]
Yuya added `coalesce()` as a built-in function with variable arity in
a2a9b7decb586, and we want to use it with more-than-two arguments in the default
`log_node` configuration in order to provide better graph node symbols for
users.

However, `test_templater` already defined this name as an alias, but with
only two parameters; this alias now conflicts with the built-in definition and
overrides it, causing the test to fail catastrophically because now the default
configuration is considered invalid.

Simply renaming it fixes this without invalidating the actual test case.

Signed-off-by: Austin Seipp <aseipp@pobox.com>
Change-Id: I18593f64b9168fc334001c2ff5d49ad8d81c006d