git: Always use OpenSSL on win32
By default, libssh2 uses Windows Cryptography Next Generation when targeting
win32. The wincng backend does not support ED25519, which is a widely-used
algorithm among git hosting services, and in some cases may be the only option
on remotes in certain configurations.

This change adds the feature `openssl-on-win32`, enabling the same one in
`libssh2-sys`, and enables it in the CI release flow, ensuring that release
builds are able to communicate with hosts using ED25519.

Fixes #3322