cli: more detailed `version` info, with short git hash
`version` now prints the short 12-byte git hash of the current commit, and this
is now prefixed with `g` to make it clear that the following characters are a
"g"it hash; we might use a `j` variant in the future, for example. The full
commit hash is still given when `--build-info` is provided.