Add a branch option to 'jj new'
Add a new option to the `jj new` command to automatically create or set a
branch. For example `jj new -m "create fix" -b new_fix` will set the branch
new_fix to the newly created commit. If the branch doesn't exist already,
it will create it automatically for the user.

The new flag is optional and can be used via the short form `-b` or the long
form `--branch`. It takes one required argument which is the branch name.