docs (direct website push): make "nightly" the preferred alias for "prerelease"
This follows up on https://github.com/martinvonz/jj/pull/3723.

As shown in the screenshot of that PR, currently the
"prerelease" version is also labeled as "main". This may
confuse people on whether it's preferred to the version
labelled "latest".

This commit changes the label to "nightly" instead of
"main".