rebase: make `-b` use `reachable()` revset
I think this is a more expected and consistent behavior. `-b` already
includes ancestors that are not the destination, as well as their
descendants. The reason we (and Mercurial) were doing just one such
back-and-forth walk was probably just that we didn't have
`reachable()` and it's not very common to notice the difference anyway
since it doesn't happen if you're already on the head commit of the
"branch"..