tests: demonstrate how `jj rebase -b` doesn't rebase "other side" of merge
When rebasing `-b` and there is a merge commit, we will rebase the
commits that are ancestors of the specific commit but not ancestors of
the destination (i.e. `<destination>..<specified>`). We also rebase
their descendants. However, we don't rebase *their* ancestors, and so
on. As shown in the updated test case, that leads to confusing
behavior where the rebased set of commits can be different if you
specify a descendant.