lib: merge `testutils` crate directly into `jj_lib::testutils`
While reading the code while working on Buck-style builds, I realized that there
is a strange cyclic dependency between `jj-lib` and `testutils`:

- `jj-lib` has a dev-dep on `testutils`
- `testutils` has a normal dep on `jj_lib`

Cargo apparently has (some?) kind of logic to topologize this sort of graph,
but it's awkward to understand or otherwise express, and has knock-on issues
with `rust-analyzer` too because certain code is only built under the proper
testing configuration, e.g. with `#[cfg(feature = "testing")]`

This instead just merges all of the `testutils` code as a single
`jj_lib::testutils` module, which breaks the cycle and also simplifies a few
of the mod imports as well. To fix the remaining compile errors we have to
explicitly use `jj_lib::testutils` first, which minimizes the overall diff. Some
call sites are simple enough to just inline though.

I do think that some of this can be pulled back out into a `jj-utils`
package that `jj-lib` unconditionally depends on and contains other various
functionality, too.

Signed-off-by: Austin Seipp <aseipp@pobox.com>