Change bg color of whitespace-only hunks
With the default color-words diff formatting, if a diff hunk (a
sequence of removed and/or added characters) consists only of whitespace
characters, it will not be visible on a terminal because the default
configuration only changes their foreground color (which has no effect
because all these characters are whitespace).

Therefore, change the background color of these diff hunks. The user
may further customize their appearance by configuring the "removed
whitespace" and "added whitespace" formats.

This was done with the help of Damian Rogers and Kyle Lippincott.