Add extra partioning for diff hunks, splitting up each slice into a section for leading whitespace, inner text, and trailing whitespace.
Previously, if a diff hunk contained both whitespace changes (leading/trailing) and text changes, the whitespace changes would not be visible within the terminal as these whitespace changes would be lumped together with text changes resulting in generic removed and added labels. But now with this extra partioning we are able to add labels to each individual section, meaning now we can style diff hunks with both text and whitespace changes.

This was done with the help of Jonathan Tan and Kyle Lippincott