conflicts: insert trailing newline when materializing conflicts
This happens for non-empty files only; the materialization code can
handle files and hunks with 0 or more lines, but each line should
terminate in a newline.

This is an imperfect fix to #3968, a better fix is TODO; see #3975.

This also implements `Merge::into_map` to avoid distracting iterations
and MergeBuilders. I could split it into a separate commit, but then
it'd have no uses.