revset: parameterize signature predicates by field
The underlying predicates now match only on the name and email,
and the revset functions construct the union of the two. These are
desired for the `.mailmap` work and aren’t currently surfaced in
the user interface, but the possibility of exposing more specific
queries in the future becomes available.