next/prev: Add config flag to control `prev/next` edit behaviour.
The flag is a tristate flag where:
  - Auto - Maintain current behaviour. This edits if
    the wc parent is not a head commit. Else, it will
    create a new commit on the parent of the wc in
    the direction of movement.
  - Always - Always edit
  - Never - Never edit, prefer the new+squash workflow.

Also add a `--no-edit` flag as the explicit inverse of `--edit` and
ensure both flags take precedence over the config.

NOTE: There is a follow up #4302 that changes the config knob to a boolean.
I'm keeping that as a separate PR for review clarity as that changes some
tests, so keeping them separate makes it easier to see what's changing.

Part of #3947