lib: inline pest grammars into `.rs` files
When building Jujutsu crates as a dependency in a Buck project, the usage of
`#[grammar]` requires the `.pest` files in the code to be in the `$PWD` relative
to where the Rust compiler is invoked on the crate; which isn't identical to how
it is done by Cargo, causing a build failure ("file not found").

Rather than special case a Buck build path into the codebase for this (only used
by downstream consumers in other projects), inlining the grammar definition into
a new Rust file makes this go away.

Signed-off-by: Austin Seipp <aseipp@pobox.com>