revsets: add a `other_working_copies()` revset
It's come up a few times that people want to mark working copies other
than the current one as immutable. We don't have a way of finding such
commits. `working_copies() ~ @` is close, but it's not correct if the
same commit is being edited in the current workspace and in another
workspace. This patch adds a `other_working_copies()` revset for it.