gitignore: add `/buck-out/` to ignore list
While working on the Buck branch, I have been bitten about a million times by
the following setup:

- Hack away
- Switch back to main
- Run `jj st` after editing a file
- `/buck-out` gets snapshotted, because it isn't ignored
- Trip max snapshot filesize error
- Pain

There are a few ways to skin the cat but this is simplest. It's an incredibly
small addition and it may be all for naught in the long run, but this will stem
the bleeding, at least.

Signed-off-by: Austin Seipp <aseipp@pobox.com>